+++ components/colorpicker/colorpicker.ts
@@ -1693,7 +1693,7 @@ export class ColorPicker
         };
         setTimeout(() =>
         {
-            document.addEventListener("mousedown", this.boundOutsideClick!);
+            document.addEventListener("pointerdown", this.boundOutsideClick!);
             document.addEventListener("keydown", this.boundEscapeKey!);
         }, 0);
     }
@@ -1703,7 +1703,7 @@ export class ColorPicker
     {
         if (this.boundOutsideClick)
         {
-            document.removeEventListener("mousedown", this.boundOutsideClick);
+            document.removeEventListener("pointerdown", this.boundOutsideClick);
             this.boundOutsideClick = null;
         }
         if (this.boundEscapeKey)